+++ net2-binding/src/main/java/org/openhab/binding/net2/Net2BindingConstants.java
@@ -45,6 +45,9 @@ public class Net2BindingConstants {
     public static final String CHANNEL_LIST_ACCESS_LEVELS = "listAccessLevels";
     public static final String CHANNEL_LIST_USERS = "listUsers";
 
+    // Security channels
+    public static final String CHANNEL_LOCKDOWN = "lockdown";
+
     public static final String CHANNEL_DOOR_CONTROL_TIMED = "controlTimed";
     // SignalR hub
     public static final String SIGNALR_HUB_PATH = "/eventHubLocal";